Add score threshold to config. #6

+++ src/main/java/de/hpi/shoprulesgenerator/properties/ShopRulesGeneratorConfig.java
@@ -16,7 +16,7 @@ import javax.validation.constraints.NotBlank;
 public class ShopRulesGeneratorConfig {
 
     @Min(0)
-    private float scoreThreshold;
+    private double scoreThreshold;
 
     @Min(0)
     private int fetchDelay;
